Split code into Loader and Property classes
